Only user without reports and investigation can be deleted

+++ Web-Applications-Architecture-and-Systems-Development/ReportSystem/Controllers/AdminController.cs
@@ -17,12 +17,18 @@ namespace ReportSystem.Controllers
         private readonly UserManager<ApplicationUser> _userManager;
         private readonly SignInManager<ApplicationUser> _signInManager;
         private readonly IAdminService _adminService;
+        private readonly IReportService _reportService;
+        private readonly IInvestigationService _investigationService;
 
-        public AdminController(UserManager<ApplicationUser> userManager, SignInManager<ApplicationUser> signInManager,IAdminService adminService)
+        public AdminController(UserManager<ApplicationUser> userManager,
+            SignInManager<ApplicationUser> signInManager,
+            IAdminService adminService, IReportService reportService, IInvestigationService investigationService)
         {
             _userManager = userManager;
             _signInManager = signInManager;
             _adminService = adminService;
+            _reportService = reportService;
+            _investigationService = investigationService;
         }
         public IActionResult Index()
         {
@@ -51,8 +57,14 @@ namespace ReportSystem.Controllers
         [HttpGet]
         public async Task<IActionResult> DeleteUser(string userId)
         {
-            await _adminService.DeleteUser(userId);
-            //_adminService.DeleteUser(userId);
+            var u= await _userManager.FindByIdAsync(userId);
+            var inv = _investigationService.CheckIfInvestigatorIsActive(u.Id);
+            var rep = _reportService.GetReportsByReporterId(u.Id).Count;
+            if (!inv && rep==0)
+            {
+                await _adminService.DeleteUser(userId);
+                return RedirectToAction("Users");
+            }
             return RedirectToAction("Users");
         }
     }
